Adds delete, get_by_id, and patch error handling

+++ server/helpers.py
@@ -2,6 +2,7 @@ from config import db
 
 from sqlalchemy.orm import joinedload
 from flask import make_response, request
+from flask_sqlalchemy import SQLAlchemy as sa
 
 
 class Routing:
@@ -28,18 +29,23 @@ class Routing:
     def delete_entry(self, id, table_class, deleted_item):
         entry = table_class.query.filter_by(id=id).first()
 
-        db.session.delete(entry)
-        db.session.commit()
-
-        response_dict = {"message": f"{deleted_item} successfully deleted"}
-
-        response = make_response(response_dict, 200)
-
-        return response
+        if not entry:
+            return make_response({'message': f'{deleted_item} not found'}, 404)
+        try:
+            db.session.delete(entry)
+            db.session.commit()
+            return make_response({'message': f"{deleted_item} successfully deleted."}, 200)
+        
+        except sa.SQLAlchemyError:
+            db.session.rollback()
+            return {'message': 'An error occurred during deletion.'}, 500
     
     def get_by_id(self, id, table_class, schema):
         entry = table_class.query.filter_by(id=id).first()
 
+        if not entry:
+            return make_response({'message': 'Entry not found'}, 404)
+        
         response = make_response(
             schema.dump(entry),
             200
@@ -50,12 +56,17 @@ class Routing:
     def patch(self, id, table_class, schema):
         entry = table_class.query.filter_by(id=id).first()
 
-        for attr in request.form:
-            setattr(entry, attr, request.form[attr])
+        if not entry:
+            return make_response({'message': 'Entry not found'}, 404)
+        try:
+            for attr in request.form:
+                setattr(entry, attr, request.form[attr])
 
-            db.session.add(entry)
-            db.session.commit()
+                db.session.add(entry)
+                db.session.commit()
 
-            response = make_response(schema.dump(entry), 200)
+                return make_response(schema.dump(entry), 200)
 
-            return response
+        except sa.SQLAlchemyError:
+            db.session.rollback()
+            return make_response({'message': 'An error occurred during update'}, 500)
